Reduced unnecesary data sent, Fixed ports, Added starter interface and more
Added a user menu for selecting application mode and to ensure the container JSON file is created if missing in DemoApplication. IntervalApplications now only processes active containers and improves actuator endpoint checks. SetupApplications now marks misconfigured containers as 'unconfigured' and provides user warnings, only if the container is not running and if the JSON does not contain the port data. ContainerClass fields are now private with appropriate getters/setters. Lastly also cleaned all of the unused imports.

Co-Authored-By: MBielefeldt1 <180258105+MBielefeldt1@users.noreply.github.com>
Co-Authored-By: dani285e <31062230+dani285e@users.noreply.github.com>